requirements.txt
Telethon==1.34.0
tweepy==4.14.0
langchain==0.1.12
chromadb==0.4.24
faiss-cpu==1.8.0
openai==1.14.2
python-dotenv==1.0.1
click==8.1.7
transformers==4.38.2
torch
accelerate==0.27.2
sentence-transformers==2.5.1 